added auth core utils moved from security module
